backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-bcrypt
pyjwt